pyproject.toml
[tool.poetry]
authors = ["hashir omer <hashiromer621@gmail.com>"]
description = "Keeping tabs on freelancing/job opportunities on reddit"
name = "jobs_searcher"
version = "0.1.0"

[tool.poetry.scripts]
start = "jobs_searcher.main:fetch_jobs"

[tool.poetry.dependencies]
asyncpraw = "^7.5.0"
dependency-injector = "^4.40.0"
dnspython = "^2.2.1"
motor = "^3.0.0"
pydantic = "^1.10.1"
pylint = "^2.15.0"
pymongo = "^4.2.0"
python = "^3.10"
python-dotenv = "^0.21.0"
pytz = "^2022.2.1"
win10toast-click = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
